Refactored app to use new actions and reducers path
